Move pfreeathome to fah subfolder to make it better testable
Also prepares splitting up library and component
